backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.1
pip~=24.3.1
pillow~=11.0.0
colorama~=0.4.6
pytz~=2024.2
qrcode~=8.0
django-rest-swagger
drf-yasg
djangorestframework-simplejwt
